memoize dresp base df
