+++ bmeg_app/db/__init__.py
@@ -4,6 +4,7 @@ import yaml
 
 import gripql
 import dash
+import pandas as pd
 
 from ..app import app
 from elasticsearch import Elasticsearch
@@ -51,3 +52,38 @@ def gene_search(query):
 @cache.memoize(timeout=TIMEOUT)
 def get_vertex_label_count(label):
     return G.query().V().hasLabel(label).count().execute()[0]['count']
+
+def line2disease(lines_list):
+    '''Dictionary mapping cell line GIDs to reported primary disease'''
+    disease_dict = {}
+    q=G.query().V(lines_list).render(["$._gid",'$._data.cellline_attributes.Primary Disease'])
+    for row in q:
+        disease_dict[row[0]]=row[1] # some vals are None
+    return disease_dict
+        
+@cache.memoize(timeout=TIMEOUT)
+def get_base_matrix(project, drug_resp,selected_drug,selected_disease):
+    '''Get row col matrix of cell line vs drug for a specific project'''   
+    q = G.query().V(project).out("cases").as_("ca").out("samples").out("aliquots").out("drug_response").as_("dr").out("compounds").as_("c")
+    q = q.render(["$ca._gid", "$c._gid", drug_resp])
+    data = {}
+    for row in q:
+        if row[0] not in data: 
+            data[row[0]] = { row[1] :  row[2] } 
+        else: 
+            data[row[0]][row[1]] = row[2]  
+    df = pd.DataFrame(data).transpose()
+    df = df.sort_index().sort_index(axis=1) #sort by rows, cols
+    # create disease mapping dict
+    disease =line2disease(list(df.index))
+    # exclude non disease related derived cell lines
+    new_col=[]
+    for a in list(df.index):
+        new_col.append(disease.get(a))
+    df['disease']=new_col
+    subset_df=df[df['disease']==selected_disease]
+    # set established drug to first col and rm disease col
+    new_col=subset_df.pop(selected_drug)
+    subset_df.insert(0, selected_drug, new_col)
+    subset_df.pop('disease')
+    return subset_df
